Update rasterwriter.py
In my ubuntu 14.04 box, the inkscape output was using "," instead of "," for the floats returned by stdout inside the exportRaste method. The output was, for example: "Area 232,857:90:834,286:422,857 exported to 601 x 333 pixels (90 dpi)". So raster export was throwing a Python error "TypeError: 'NoneType' object is not iterable". I solved this by replacing "," with ".", and now it works like a charm.